+++ geex-models/src/main/java/nl/tudelft/context/model/graph/GraphParser.java
@@ -1,8 +1,6 @@
-package nl.tudelft.context.model.graph.parser;
+package nl.tudelft.context.model.graph;
 
 import nl.tudelft.context.model.Parser;
-import nl.tudelft.context.model.graph.GraphMap;
-import nl.tudelft.context.model.graph.Node;
 
 import java.io.BufferedReader;
 import java.util.ArrayList;
