+++ geex-models/src/main/java/nl/tudelft/context/model/graph/NodeParser.java
@@ -1,6 +1,4 @@
-package nl.tudelft.context.model.graph.parser;
-
-import nl.tudelft.context.model.graph.Node;
+package nl.tudelft.context.model.graph;
 
 import java.util.Arrays;
 import java.util.HashSet;
