+++ geex-models/src/test/java/nl/tudelft/context/model/graph/StackGraphTest.java
@@ -1,7 +1,6 @@
 package nl.tudelft.context.model.graph;
 
 import de.saxsys.javafx.test.JfxRunner;
-import nl.tudelft.context.model.graph.parser.GraphParser;
 import nl.tudelft.context.service.LoadService;
 import org.junit.Test;
 import org.junit.runner.RunWith;
